requirements.txt
# Core
langchain
langchain-anthropic
langgraph
anthropic
python-dotenv

# Visualization
pyecharts

# Web UI
streamlit

# Data
pandas